+++ CHANGELOG.md
@@ -59,7 +59,7 @@
 - **algorithm/math:** add game ([20dfd5c](https://github.com/air-supply94/js-algorithm/commit/20dfd5c1c6c15e47220442bcba3d0393bfcbadc9))
 - **algorithm/math:** add getRandom ([da6ffd9](https://github.com/air-supply94/js-algorithm/commit/da6ffd976e772de2d5b6b6903ec597a752391bf1))
 - **algorithm/math:** add missingNumber ([168e330](https://github.com/air-supply94/js-algorithm/commit/168e33009a3ad21ba4a15c56c65367588e6c03ad))
-- **algorithm/math:** add subarraySum ([8f23792](https://github.com/air-supply94/js-algorithm/commit/8f23792c76acb6f8f2d60d9e6ed24a117f14e4ce))
+- **algorithm/math:** add subarraySum ([69730c2](https://github.com/air-supply94/js-algorithm/commit/69730c229b2b325e902a04ce93692e72bd0d14e4))
 - **algorithm/math:** add trailingZeroes ([cefab4c](https://github.com/air-supply94/js-algorithm/commit/cefab4c03594616e8c8ff8ed447a4bdd1febff3d))
 - **algorithm/math:** add WSRARes ([d6589c4](https://github.com/air-supply94/js-algorithm/commit/d6589c43f9c6dd469b0a642b4df9f43a136de3b8))
 - **algorithm/slidingWindow:** add checkInclusion ([e952fc8](https://github.com/air-supply94/js-algorithm/commit/e952fc852e3392f5a1e92fdbb6cd1fa02d20d8d7))